move lsq activation quantizers from qat to trainable infra
